Rename pcpairing to pairing package
